requirements.txt
# Django Framework
Django==5.2
python-dotenv

# LangChain and AI Libraries
langchain
langchain-openai
langchain-community
langchain-core
openai

# Vector Store and Embeddings
faiss-cpu

# PDF Processing
pypdf

# Development and Debugging
python-decouple

# Optional: For production deployment
gunicorn
whitenoise

# Optional: For testing
pytest
pytest-django
